Fix dependency rule issue with common types
Jira ticket: CAMS-421
