+++ common/src/queue/dataflow-types.ts
@@ -1,5 +1,4 @@
 import { DxtrCase } from '../cams/cases';
-import { CamsError } from '../../../backend/lib/common-errors/cams-error';
 
 export type CaseSyncEvent = {
   type: 'CASE_CHANGED' | 'MIGRATION';
@@ -36,42 +35,5 @@ export function getDefaultSummary(
     completed: 0,
     faulted: 0,
     ...override,
-  };
-}
-
-export type MaybeError<E extends CamsError = CamsError> = {
-  error?: E;
-};
-
-export type MaybeData<T = unknown> = MaybeError & {
-  data?: T;
-};
-
-export type MaybeVoid = MaybeError & {
-  success?: true;
-};
-
-export type MaybeCaseSyncEvents = MaybeError & {
-  events?: CaseSyncEvent[];
-};
-
-export type QueueError<E extends CamsError = CamsError> = {
-  type: 'QUEUE_ERROR';
-  module: string;
-  // NOTE: James says this is gonna bite us when we change the activityName. He is probably right. -- BTP
-  activityName: string;
-  error: E;
-};
-
-export function buildQueueError(
-  error: CamsError,
-  module: string,
-  activityName: string,
-): QueueError {
-  return {
-    type: 'QUEUE_ERROR',
-    module,
-    activityName,
-    error,
   };
 }
